feat: separate icons and logos
