+++ src/components/Icons/NodeIcon.tsx
