+++ src/components/Icons/PairIcon.tsx
