+++ src/pages/ConnectDevicePage/ConnectDevicePage.tsx
@@ -3,12 +3,12 @@ import BreadcrumbBar from '../../components/BreadcrumbBar'
 import { Button as StatusButton, Tag, Text, Avatar, Checkbox } from '@status-im/components'
 import { Label, Separator, XStack, YStack } from 'tamagui'
 import PageWrapperShadow from '../../components/PageWrapperShadow'
-import NimbusLogo from '../../components/NimbusLogo'
+import NimbusLogo from '../../components/Logos/NimbusLogo'
 import Titles from '../../components/Titles'
-import NodeIcon from '../../components/NodeIcon'
-import ConnectIcon from '../../components/ConnectIcon'
-import PairIcon from '../../components/PairIcon'
-import CreateIcon from '../../components/CreateIcon'
+import NodeIcon from '../../components/Icons/NodeIcon'
+import ConnectIcon from '../../components/Icons/ConnectIcon'
+import PairIcon from '../../components/Icons/PairIcon'
+import CreateIcon from '../../components/Icons/CreateIcon'
 import LabelInputField from '../../components/LabelInputField'
 
 const ConnectDevicePage = () => {
